Refactored image upload and added dispaly
